final changes suggested by github actions
