+++ tests/pdf_viewer/conftest.py
@@ -1,8 +1,13 @@
+import os
+import time
 from shutil import copyfile
 
 import pytest
 
 from modules.page_object import GenericPdf
+
+DOWNLOAD_TIMEOUT_SEC = 5.0
+POLL_INTERVAL_SEC = 1.0
 
 
 @pytest.fixture()
@@ -38,3 +43,26 @@ def pdf_file_path(tmp_path, file_name: str):
 @pytest.fixture()
 def pdf_viewer(driver, pdf_file_path):
     return GenericPdf(driver, pdf_url=f"file://{pdf_file_path}")
+
+
+@pytest.fixture()
+def wait_for_file_download():
+    """Return a helper that blocks until a file finishes downloading."""
+
+    def _wait_for_file_download(
+        saved_pdf_path, timeout=DOWNLOAD_TIMEOUT_SEC, interval=POLL_INTERVAL_SEC
+    ) -> bool:
+        """Wait until file exists on disk or raise a pytest failure."""
+        start_time = time.time()
+        while time.time() - start_time < timeout:
+            if os.path.exists(saved_pdf_path):
+                initial_size = os.path.getsize(saved_pdf_path)
+                time.sleep(interval)
+                final_size = os.path.getsize(saved_pdf_path)
+                if initial_size == final_size and final_size > 0:
+                    return True
+            time.sleep(interval)
+        pytest.fail(f"The file was not downloaded within {timeout:.1f} seconds.")
+        return None
+
+    return _wait_for_file_download
