+++ tests/pdf_viewer/test_download_pdf_with_form_fields.py
@@ -8,29 +8,6 @@ from modules.page_object_generics import GenericPdf
 
 PDF_FILE_NAME = "i-9.pdf"
 DOWNLOADED_PDF_REGEX = r"i-9.*\.pdf"
-DOWNLOAD_TIMEOUT_SEC = 5.0
-POLL_INTERVAL_SEC = 1.0
-
-
-def _wait_for_file_download(
-    saved_pdf_path, timeout=DOWNLOAD_TIMEOUT_SEC, interval=POLL_INTERVAL_SEC
-) -> None:
-    """Wait until file exists on disk or raise a pytest failure."""
-
-    start_time = time.time()
-    while time.time() - start_time < timeout:
-        if os.path.exists(saved_pdf_path):
-            initial_size = os.path.getsize(saved_pdf_path)
-            time.sleep(interval)
-            final_size = os.path.getsize(saved_pdf_path)
-
-            if initial_size == final_size and final_size > 0:
-                return True
-
-        time.sleep(interval)
-
-    pytest.fail(f"The file was not downloaded within {timeout:.1f} seconds.")
-    return None
 
 
 @pytest.fixture()
@@ -57,6 +34,7 @@ def test_download_pdf_with_form_fields(
     delete_files,
     downloads_folder: str,
     delete_files_regex_string,
+    wait_for_file_download,
 ):
     """
     C1020326 Download pdf with form fields
@@ -81,7 +59,7 @@ def test_download_pdf_with_form_fields(
     saved_pdf_path = os.path.join(downloads_folder, PDF_FILE_NAME)
 
     # Verify if the file exists
-    _wait_for_file_download(saved_pdf_path)
+    wait_for_file_download(saved_pdf_path)
 
     # Open the saved pdf and check if the edited field is displayed
     driver.get("file://" + os.path.realpath(saved_pdf_path))
